Fix selection rendering bug
