+++ pyspinw/gui/rendering/shaders/object_vertex.glsl
@@ -11,9 +11,8 @@ out vec3 Normal;
 
 void main()
 {
-    FragPos = vec3(model * vec4(aPos, 1.0));
-    Normal = mat3(transpose(inverse(model))) * aNormal;
+    FragPos = vec3(model * vec4(aPos, 1.0)); // Position in world
+    Normal = mat3(transpose(inverse(model))) * aNormal; // Normal
 
     gl_Position = projectionView * vec4(FragPos, 1.0);
-//    gl_Position = projectionView * vec4(aPos, 1.0);
 }
